prov/efa: extract efa_rdm_pke_rtr.c from rxr_pkt_type_req.c
Move RTR related fucntions to efa_rdm_pke_rtr.c and
changed the prefix from rxr_pkt to efa_rdm_pke.
